+++ modules_k/presence_mwi/add_events.c
@@ -99,7 +99,7 @@ int mwi_publ_handl(struct sip_msg* msg)
 
 err:
     LM_ERR("check of body <%.*s> failed at character number %d\n",
-	   body.len, body.s, at - body.s + 1);
+	   body.len, body.s, (int)(at - body.s + 1));
     return -1;
 
 }
